set qid.vers (Erik Quanstrom)

+++ src/lib9/fmt/fmtfdflush.c
@@ -27,7 +27,7 @@ __fmtFdFlush(Fmt *f)
 	int n;
 
 	n = (char*)f->to - (char*)f->start;
-	if(n && write((int)f->farg, f->start, n) != n)
+	if(n && write((uintptr)f->farg, f->start, n) != n)
 		return 0;
 	f->to = f->start;
 	return 1;
