+++ src/lib9/fmt/runevsmprint.c
@@ -36,7 +36,7 @@ runeFmtStrFlush(Fmt *f)
 
 	if(f->start == nil)
 		return 0;
-	n = (int)f->farg;
+	n = (uintptr)f->farg;
 	n *= 2;
 	s = (Rune*)f->start;
 	f->start = realloc(s, sizeof(Rune)*n);
@@ -47,7 +47,7 @@ runeFmtStrFlush(Fmt *f)
 		free(s);
 		return 0;
 	}
-	f->farg = (void*)n;
+	f->farg = (void*)(uintptr)n;
 	f->to = (Rune*)f->start + ((Rune*)f->to - s);
 	f->stop = (Rune*)f->start + n - 1;
 	return 1;
@@ -67,7 +67,7 @@ runefmtstrinit(Fmt *f)
 	f->to = f->start;
 	f->stop = (Rune*)f->start + n - 1;
 	f->flush = runeFmtStrFlush;
-	f->farg = (void*)n;
+	f->farg = (void*)(uintptr)n;
 	f->nfmt = 0;
 	return 0;
 }
